refactor(cli): implement BusyBox pattern to eliminate Go runtime duplication
Refactored CLI from "Embedded Binaries" to "BusyBox Pattern" to eliminate
redundant Go runtime duplication and reduce binary size.

Changes:
- Refactored stigmer-server to importable library (pkg/server/server.go)
- Refactored workflow-runner to importable library (pkg/runner/runner.go)
- Added hidden CLI commands (internal-server, internal-workflow-runner)
- Updated daemon to spawn CLI itself with hidden commands (BusyBox pattern)
- Simplified embedded binaries (only agent-runner Python binary)

Results:
- CLI size: 150MB → 126MB (24MB reduction, 16% smaller)
- Go runtime copies: 3 → 1 (shared across components)
- Build process: 3 Go builds → 1 Go build (simplified)
- User experience: No changes (transparent optimization)

Technical details:
- stigmer-server: main.go (352 lines) → 13 lines + pkg/server/server.go (393 lines)
- workflow-runner: package main → package worker (importable)
- Daemon spawns: `stigmer internal-server` instead of extracted binary
- Only agent-runner (Python/PyInstaller) remains as embedded binary

Benefits:
- Faster downloads (smaller binary)
- Simpler CI/CD (fewer build steps)
- Easier debugging (single binary)
- Consistent Go runtime version

Following Gemini's architectural recommendation to treat own Go code as
importable libraries rather than "foreign" binaries.

Refs: _changelog/2026-01/2026-01-21-074915-busybox-pattern-cli-refactoring.md
Refs: _cursor/seperate-binnary.md (Gemini conversation)
