+++ backend/services/workflow-runner/cmd/worker/root.go
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package main
+package worker
 
 import (
 	"context"
