+++ client-apps/cli/embedded/embedded_darwin_arm64.go
@@ -7,25 +7,11 @@ import (
 )
 
 // Darwin ARM64 (Apple Silicon) embedded binaries
+// Note: Only agent-runner is embedded (Python binary)
+// stigmer-server and workflow-runner are compiled into the CLI (BusyBox pattern)
 //
-//go:embed binaries/darwin_arm64/stigmer-server
-var stigmerServerBinary []byte
-
-//go:embed binaries/darwin_arm64/workflow-runner
-var workflowRunnerBinary []byte
-
 //go:embed binaries/darwin_arm64/agent-runner
 var agentRunnerBinary []byte
-
-// GetStigmerServerBinary returns the embedded stigmer-server binary for darwin/arm64
-func GetStigmerServerBinary() ([]byte, error) {
-	return stigmerServerBinary, nil
-}
-
-// GetWorkflowRunnerBinary returns the embedded workflow-runner binary for darwin/arm64
-func GetWorkflowRunnerBinary() ([]byte, error) {
-	return workflowRunnerBinary, nil
-}
 
 // GetAgentRunnerBinary returns the embedded agent-runner binary for darwin/arm64
 func GetAgentRunnerBinary() ([]byte, error) {
